Code cleaned up.

+++ EnderDeath/src/pl/socialCraft/enderdeath/commands/TimeCommand.java
@@ -10,7 +10,8 @@ public class TimeCommand extends Command {
 	@Override
 	public void performCommand(Player sender, String cmd, String[] args) {
 		if (EnderDeath.getRound().getPlayerTeam(sender) != null)
-			sender.sendMessage(Config.getMessage("time", formatTime(EnderDeath.getRound().getTime())));
+			sender.sendMessage(Config.getMessage("time", formatTime(EnderDeath
+					.getRound().getTime())));
 	}
 	
 	@Override
@@ -18,8 +19,8 @@ public class TimeCommand extends Command {
 		return Config.getCommand("time");
 	}
 	
-	private String formatTime(int time){
-		return String.valueOf((int)time / 60) + ":" + String.valueOf(time % 60);
+	private String formatTime(int time) {
+		return String.valueOf(time / 60) + ":" + String.valueOf(time % 60);
 	}
 	
 }
